스킬오브젝트 비활성화 조건 수정

+++ LevelManager.cpp
@@ -143,7 +143,7 @@ void LevelManager::DeleteObject(GameObject& me)
 	assert(curLevel);
 
 	auto& objects = curLevel->objects[(int)me.GetLayer()];
-	int size = objects.size();
+	size_t size = objects.size();
 
 	for (int i=0; i<size; ++i)
 	{
